Refactoring simulation methods
