+++ SpiceSharp/Simulations/Transient/IntegrationMethods/Breakpoints.cs
