+++ SpiceSharp/Simulations/Transient/IntegrationMethods/Trapezoidal.cs
